requirements.txt
cx_Oracle==8.3.0
fastapi==0.111.0
python-dotenv==1.0.1
uvicorn==0.30.1